rag_project/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-policy-bot"
version = "0.1.0"
description = "RAG Policy Analysis Bot"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "python-dotenv",
    "langchain-community",
    "langchain-core",
    "openai",
    "faiss-cpu",
    "scikit-learn",
    "numpy",
    "tiktoken"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio"
]

[tool.setuptools]
packages = ["src"]
